Rename an autogenerated file.
Change the name of the autogenerated header file containing the "typedef
struct" declarations for all classes within a Clownfish parcel from "boil.h"
(a legacy of the Clownfish's old name, "Boilerplater"), to "parcel.h".  It's
still not ideal, as it should be named after the parcel itself, but we'll
cross that bridge later.

git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/trunk@1143514 13f79535-47bb-0310-9956-ffa450edef68