Fix refcounting of args in callbacks.
When compiled Clownfish code manipulates a host subclass and calls back into
the host language, refcounts for return values were being handled correctly,
but refcounts for arguments were not.  This is because the Host_callback_xxx
routines have a particular signature: Host_callback_obj returns an incremented
Obj, but other than that they are refcount neutral.

This patch adjusts the override wrappers which call back into the host
language to ensure that their refcounting behavior matches that of the methods
they are overriding.


git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/trunk@1230857 13f79535-47bb-0310-9956-ffa450edef68