Hack to fix 'for' scoping under MSVC6
MSVC6 doesn't allow redefinition of iteration variables in 'for'
statements. This ugly hack redefines 'for' as 'if(0);else for' to make
this work. For now, this is applied to all versions of MSVC but it
shouldn't hurt.

Since the macro contains a space character, some additional quoting is
needed. The extra_compiler_flags also have to be passed as an array.

Clean up setting of compiler flags along the way.