Use charmonizer's EXTRA_CFLAGS in Perl bindings.
Start feeding the EXTRA_CFLAGS key generated by charmonizer, and migrate
the tests for LUCY_DEBUG and LUCY_VALGRIND to set extra compiler flags
from the Perl bindings to charmonizer.c.

As of this commit, flags which move to Charmonizer are only used when
compiling .c files from the `clownfish_params => source` dirs, and are
no longer supplied when compiling .xs binding files.