Additional locations for Clownfish include dirs
The environment variable CLOWNFISH_INCLUDE can be set to a
colon-separated list of directories which cfc will use as include
directories. For now, this feature is only available for the C bindings.

The C bindings will also add the following system-wide include
directories:

    * /usr/local/share/clownfish/include
    * /usr/share/clownfish/include

If CLOWNFISH_INCLUDE is set, the system-wide directories will not be used
to avoid errors when a parcel is found in multiple locations.