Determine Clownfish include dirs at build time
Previously, the list of Clownfish include dirs was compiled at
configure time when constructing the Module::Build object. M::B then
serializes the object and reloads it for subsequent build actions. As
a consequence, dependencies that were installed after configuration
weren't found. This caused CPAN installs and smokers to break.

Creating the list on demand should fix this.