Move Python files under `src`.
Moving the Python source directory into `src` means that the files are
no longer findable by python running in the python/ directory because
the current directory is on sys.path.  In later versions of Python, this
becomes an issue because the source dir (which does not contain any
shared objects) shadows the build directory (where shared objects get put).