Use PERL_NO_GET_CONTEXT
Pass the Perl interpreter context explicitly. This avoids many unneeded
calls to Perl_get_context for threaded Perls. See section "How multiple
interpreters and concurrency are supported" in the perlguts man page.

Decreases the size of the i386 binary by 15 KB which illustrates the
extent of this optimization.

Fixes CLOWNFISH-29.