Eliminate StackString
Use normal stack-allocated Strings instead of StackStrings.

The only downside is that there will be an invalid free instead of an
exception if someone tries to decref a stack String. But decref is
inherently dangerous. With heap-allocated objects, an unmatched decref
also results in a use-after-free which can't be guarded against.