Don't over-optimize final methods.
It is only valid to resolve a final method to a specific implementing
function when that function resides in the same parcel as the invocant.
A ancestor class in another parcel has the option of adding or removing
an implementation, so the descendant class cannot know for sure what
function it should resolve to.