Ensure proper Go wrapper for return vals.
Access to the capabilities of a Clownfish object wrapped in a Go object
depends on the Go wrapper being class-specific.  For instance, it does
not suffice to wrap a `*C.cfish_Hash` inside a `clownfish.ObjIMP`,
because you cannot invoke Hash operations on it.

Therefore, use WRAPAny for return values, so that the most specific
wrapper struct gets chosen.