Fix a missing refcount in the Go host error trapping.
Err_trap returns an `incremented` Err object, so we need to return an
incremented reference rather than return the pointer extracted from the
Err Go wrapper directly.  If we don't do that, we're arming a time bomb
which will go off the first GC after the caller disposes of their
refcount.