Ignore classes from outdated .cfh files
If Clownfish 0.5 is installed over Clownfish 0.4, a couple of old .cfh
files like VArray.cfh and LockFreeRegistry.cfh remain. These files are
processed by CFC which means that C headers are created for them, for
example. This is mostly harmless and will be fixed later when we start
to install .cfh files into separate directories. The only real
showstopper is that the build breaks when examining the Clownfish binary
for host-specific data like method aliases. This is fixed by ignoring
classes that aren't found in the binary.

Fixes CLOWNFISH-89.