Improve error message in case of invalid UTF-8
Add a new string helper StrHelp_validate_utf8 and a corresponding
macro that validates a UTF-8 string and throws an error if invalid
UTF-8 is encountered. The error message shows where the invalid UTF-8
was found and prints the offending bytes as hex.

Obsoletes the old hack that wrote an error message directly to stderr.

First part of CLOWNFISH-97.