Implement Clownfish::Test::run_tests in Perl
The XS implementation needed headers from the Clownfish parcel. If we
decide to move the tests to a separate binary, this would require to
move the implementation to a separate C file (or add a mechanism to
include custom headers from the generated .xs file).

Implementing this function in Perl is solves this problem.