Perl build with make
Let Charmonizer create a Makefile for the Perl bindings that builds the
core object files. The list of object files is passed in a Charmonizer
variable named `CORE_OBJECTS`.

Make options can be passed by running

    perl Build.PL --charmonizer_params make_options=<options>

Module::Build options can also be specified in ~/.modulebuildrc, so an
easy way to always launch parallel builds is to add a line like

    Build_PL --charmonizer_params make_options=-j5