Add "xs_prereqs" module parameter
This parameter lists the names of other Clownfish XS modules whose
symbols are referenced from a module. The corresponding libraries
are linked against on Windows.

This obviates the need to call cf_linker_flags manually.

If we start to install parcel metadata, it should be possible to
determine the XS prerequisites automatically (see CLOWNFISH-83).