Fix shared library compilation for Python
Untested. AFAIU, the static library built for Python is linked into
a shared library later. This requires to use -fPIC and define
CFP_CFISH.

I think it would be better if the Python bindings followed the
approach I took for Perl: Pass a list of object files instead of
using a static library.