New directory layout for installed Clownfish headers
The .cfh files of a parcel are installed to a separate directory:

    $prefix/$parcel/$version

The installed .cfp file is renamed to "parcel.json".

This makes it possible to parse only those .cfh files that are really
required. Also, .cfh files for different versions of a parcel can be
installed. A couple of checks for clashes between files and classes
from source and include directories become unnecessary.

For the C bindings, write headers for installed parcels to the autogen
directory first.

Enable major version checks for parcels from include directories.