Sort class arrays of parcels
Some of the code generators require that parent classes are processed
before subclasses. Sort the class array accordingly when building the
hierarchy.

Also sort subclasses by class name to guarantee a consistent order.
This should make the CFC output completely deterministic. (Except for
output that depends on the order of parcels in the global parcel
registry. I think the only example is the Perl typemap.)